Using prtconf to gather most data
git-svn-id: https://xvr2.svn.sourceforge.net/svnroot/xvr2@520 07303567-df1b-0410-b539-9fd000affbe4

+++ trunk/libs/xvr2platform/xvr2/ProcessRunner.cpp
@@ -227,7 +227,7 @@ namespace xvr2 {
 			_opaque *o = (_opaque *)stack;
 			int retcode;
 #ifdef USE_DEBUG
-			debugConsole << "Waiting for process: " << o->pid << " to finish..." << NL;
+			debugConsole << "Waiting for process: " << (UInt32)o->pid << " to finish..." << NL;
 #endif
 			if(waitpid(o->pid, &retcode, 0) == -1){
 				throw ProcessException(errno, "Weird problem on ProcessRunner::wait()");
